# payments/internal/rest/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Payments API",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "PaymentsService"
    }
  ],
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/payments": {
      "post": {
        "summary": "Authorize a future payment",
        "operationId": "authorizePayment",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentspbAuthorizePaymentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/paymentspbAuthorizePaymentRequest"
            }
          }
        ],
        "tags": [
          "Payment"
        ]
      }
    },
    "/api/payments/invoices": {
      "get": {
        "summary": "get invoice list",
        "operationId": "getInvoices",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentspbGetInvoicesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Invoice"
        ]
      },
      "post": {
        "summary": "Create an invoice for payment",
        "operationId": "createInvoice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentspbCreateInvoiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/paymentspbCreateInvoiceRequest"
            }
          }
        ],
        "tags": [
          "Invoice"
        ]
      }
    },
    "/api/payments/invoices/{id}/pay": {
      "put": {
        "summary": "Pay an invoice",
        "operationId": "payInvoice",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentspbPayInvoiceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PaymentsServicePayInvoiceBody"
            }
          }
        ],
        "tags": [
          "Invoice"
        ]
      }
    },
    "/api/payments/payments": {
      "get": {
        "summary": "get Payment list",
        "operationId": "getPayments",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/paymentspbGetPaymentsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Payment"
        ]
      }
    }
  },
  "definitions": {
    "PaymentsServicePayInvoiceBody": {
      "type": "object"
    },
    "paymentspbAdjustInvoiceResponse": {
      "type": "object"
    },
    "paymentspbAuthorizePaymentRequest": {
      "type": "object",
      "properties": {
        "customerId": {
          "type": "string"
        },
        "amount": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "paymentspbAuthorizePaymentResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "paymentspbCancelInvoiceResponse": {
      "type": "object"
    },
    "paymentspbConfirmPaymentResponse": {
      "type": "object"
    },
    "paymentspbCreateInvoiceRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string"
        },
        "paymentId": {
          "type": "string"
        },
        "amount": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "paymentspbCreateInvoiceResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "paymentspbGetInvoicesResponse": {
      "type": "object",
      "properties": {
        "invoices": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/paymentspbInvoice"
          }
        }
      }
    },
    "paymentspbGetPaymentsResponse": {
      "type": "object",
      "properties": {
        "payments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/paymentspbPayment"
          }
        }
      }
    },
    "paymentspbInvoice": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "orderId": {
          "type": "string"
        },
        "amount": {
          "type": "number",
          "format": "double"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "paymentspbPayInvoiceResponse": {
      "type": "object"
    },
    "paymentspbPayment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "customerId": {
          "type": "string"
        },
        "amount": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}